Migrate custom behavior in existing controllers to new location
