+++ src/Controller/API/IngestionException.php
@@ -6,7 +6,7 @@ namespace App\Controller\API;
 
 use App\Entity\Manager\IngestionExceptionManager;
 
-class IngestionException extends ReadWriteController
+class IngestionException extends ReadOnlyController
 {
     public function __construct(IngestionExceptionManager $manager)
     {
