+++ src/Controller/API/MeshConcept.php
@@ -6,7 +6,7 @@ namespace App\Controller\API;
 
 use App\Entity\Manager\MeshConceptManager;
 
-class MeshConcept extends ReadWriteController
+class MeshConcept extends ReadOnlyController
 {
     public function __construct(MeshConceptManager $manager)
     {
