+++ src/Controller/API/MeshPreviousIndexing.php
@@ -6,7 +6,7 @@ namespace App\Controller\API;
 
 use App\Entity\Manager\MeshPreviousIndexingManager;
 
-class MeshPreviousIndexing extends ReadWriteController
+class MeshPreviousIndexing extends ReadOnlyController
 {
     public function __construct(MeshPreviousIndexingManager $manager)
     {
