+++ src/Controller/API/User.php
@@ -4,12 +4,152 @@ declare(strict_types=1);
 
 namespace App\Controller\API;
 
+use App\Classes\SessionUserInterface;
 use App\Entity\Manager\UserManager;
+use App\RelationshipVoter\AbstractVoter;
+use App\Service\ApiRequestParser;
+use App\Service\ApiResponseBuilder;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\HttpException;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Component\Serializer\Serializer;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
+
+/**
+ * Class UsersController
+ * We have to handle a special 'q' parameter
+ * as well as special handling for ICS feed keys
+ * so users needs its own controller
+ */
 
 class User extends ReadWriteController
 {
+    /**
+     * @var UserManager
+     */
+    protected $manager;
+
     public function __construct(UserManager $manager)
     {
         parent::__construct($manager, 'users');
+    }
+
+    /**
+     * Handle the special 'q' parameter for courses
+     * @internal $manager UserM
+     */
+    public function getAll(
+        string $version,
+        Request $request,
+        AuthorizationCheckerInterface $authorizationChecker,
+        ApiResponseBuilder $builder
+    ): Response {
+        $q = $request->get('q');
+        $parameters = ApiRequestParser::extractParameters($request);
+
+        if (null !== $q && '' !== $q) {
+            $dtos = $this->manager->findUserDTOsByQ(
+                $q,
+                $parameters['orderBy'],
+                $parameters['limit'],
+                $parameters['offset'],
+                $parameters['criteria'],
+            );
+
+            $filteredResults = array_filter($dtos, function ($object) use ($authorizationChecker) {
+                return $authorizationChecker->isGranted(AbstractVoter::VIEW, $object);
+            });
+
+            //Re-index numerically index the array
+            $values = array_values($filteredResults);
+
+            return $builder->buildPluralResponse($this->endpoint, $values, Response::HTTP_OK);
+        }
+
+        return parent::getAll($version, $request, $authorizationChecker, $builder);
+    }
+
+    /**
+     * When Users are submitted with an empty icsFeedKey value that overrides
+     * the created key.  This happens when new users are created and they don't have a
+     * key yet.  Instead of using the blank key we need to keep the one that is generated
+     * in the User entity constructor.
+     */
+    public function postUser(
+        string $version,
+        Request $request,
+        ApiRequestParser $requestParser,
+        ValidatorInterface $validator,
+        AuthorizationCheckerInterface $authorizationChecker,
+        ApiResponseBuilder $builder,
+        Serializer $serializer
+    ) {
+        $data = $requestParser->extractPostDataFromRequest($request, $this->endpoint);
+        $dataWithoutEmptyIcsFeed = array_map(function ($obj) {
+            if (is_object($obj) && property_exists($obj, 'icsFeedKey')) {
+                if (empty($obj->icsFeedKey)) {
+                    unset($obj->icsFeedKey);
+                }
+            }
+
+            return $obj;
+        }, $data);
+
+        $class = $this->manager->getClass() . '[]';
+        $json = json_encode($dataWithoutEmptyIcsFeed);
+        $entities = $serializer->deserialize($json, $class, 'json');
+
+        foreach ($entities as $entity) {
+            $errors = $validator->validate($entity);
+            if (count($errors) > 0) {
+                $errorsString = (string) $errors;
+
+                throw new HttpException(Response::HTTP_BAD_REQUEST, $errorsString);
+            }
+            if (! $authorizationChecker->isGranted(AbstractVoter::CREATE, $entity)) {
+                throw new AccessDeniedException('Unauthorized access!');
+            }
+        }
+
+        foreach ($entities as $entity) {
+            $this->manager->update($entity, false);
+        }
+        $this->manager->flush();
+
+        return $builder->buildPluralResponse($this->endpoint, $entities, Response::HTTP_CREATED);
+    }
+
+    /**
+     * Only a root user can make other users root.
+     * This has to be done here because by the time it reaches the voter the
+     * current user object in the session has been modified
+     */
+    public function putUser(
+        string $version,
+        string $id,
+        Request $request,
+        ApiRequestParser $requestParser,
+        ValidatorInterface $validator,
+        AuthorizationCheckerInterface $authorizationChecker,
+        ApiResponseBuilder $builder,
+        TokenStorageInterface $tokenStorage
+    ) {
+        $entity = $this->manager->findOneBy(['id' => $id]);
+        if ($entity) {
+            $obj = $requestParser->extractPutDataFromRequest($request, $this->endpoint);
+            /** @var SessionUserInterface $sessionUser */
+            $sessionUser = $tokenStorage->getToken()->getUser();
+            if (
+                $obj->root &&
+                (!$sessionUser->isRoot() && !$entity->isRoot())
+            ) {
+                throw new AccessDeniedException('Unauthorized access!');
+            }
+        }
+
+        return parent::put($version, $id, $request, $requestParser, $validator, $authorizationChecker, $builder);
     }
 }
