--- src/Controller/ApiControllerInterface.php
@@ -1,18 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Controller;
-
-/**
- * Interface ApiControllerInterface
- *
- * This is used to tag controllers which need to be
- * authorized against a white list of possible endpoints in
- * ValidApiObjectListener
- *
- */
-interface ApiControllerInterface
-{
-
-}
