--- src/Controller/MeshDescriptorController.php
@@ -1,77 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Controller;
-
-use App\Entity\Manager\MeshDescriptorManager;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\GoneHttpException;
-
-/**
- * Class MeshDescriptorController
- * We have to handle a special 'q' parameter on meshDescriptors
- * so it needs its own controller
- */
-class MeshDescriptorController extends ApiController
-{
-    /**
-     * Handle 'q' parameter in queries
-     * @inheritdoc
-     */
-    public function getAllAction($version, $object, Request $request)
-    {
-        $q = $request->get('q');
-        $parameters = $this->extractParameters($request);
-
-        /** @var MeshDescriptorManager $manager */
-        $manager = $this->getManager($object);
-
-        if (null !== $q) {
-            $result = $manager->findMeshDescriptorsByQ(
-                $q,
-                $parameters['orderBy'],
-                $parameters['limit'],
-                $parameters['offset']
-            );
-
-            return $this->resultsToResponse($result, $this->getPluralResponseKey($object), Response::HTTP_OK);
-        }
-
-        return parent::getAllAction($version, $object, $request);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function postAction($version, $object, Request $request)
-    {
-        $this->fourTenAction();
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function putAction($version, $object, $id, Request $request)
-    {
-        $this->fourTenAction();
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function deleteAction($version, $object, $id)
-    {
-        $this->fourTenAction();
-    }
-
-    /**
-     * Generic action used by the router to send a 410 GONE
-     * to anyone trying to POST, PUT or DELETE a MeSH Descriptor
-     */
-    public function fourTenAction()
-    {
-        throw new GoneHttpException('Creating, updating and deleting MeSH Descriptors is no longer supported.');
-    }
-}
